// app/lib/extensions/interfaces.ts
export interface PackageFiles {
    [path: string]: FileDeclaration;
}

/**
 * Describes a file in a manifest
 */
export interface FileDeclaration {
    assetType?: string;
    contentType?: string;
    auto?: boolean;
    path: string;
    content?: string;
    partName?: string;
    lang?: string;
    addressable?: boolean;
}
// ...
export interface ResourcesFile {
    [key: string]: string;
}